requirements.txt
streamlit>=1.32.0
google-genai>=1.0.0
ebooklib>=0.18
beautifulsoup4>=4.12.0
pdfplumber>=0.10.0
python-docx>=1.1.0
lxml>=5.0.0
mobi>=0.3.3
